JAMES-2993 Package relocation
